add s3 service for user

+++ ms-userservice/src/main/java/com/fundoonote/msuserservice/repositories/UserDAO.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.fundoonote.msuserservice.models.User;
 
-public interface UserDAO extends CrudRepository<User, Long> {
+public interface UserDAO extends CrudRepository<User, Integer> {
 
 	User findByEmail(String email);
 }
